requirements.txt
faster-whisper>=1.0.0
llama-cpp-python>=0.2.75
tqdm>=4.66.0
